move profile out of commands and into cog with slashgroup
