+++ charts/agimus/Chart.yaml
@@ -2,5 +2,5 @@ apiVersion: v2
 name: agimus
 description: A helm chart for a discord bot that also runs a mysql db
 type: application
-version: v1.8.16
-appVersion: v1.8.16
+version: v1.9.1
+appVersion: v1.9.1
