+++ data/help/default.txt
@@ -1,15 +1,18 @@
 
-*AGIMUS is a multipurpose Discord bot by the Friends of Desoto! AGIMUS does all sorts of things, including: XP/leveling, currency, trivia, poker, slots, AI chat, and more!*
+*AGIMUS is a multipurpose Discord bot by the Friends of DeSoto! AGIMUS does all sorts of things, including: XP/leveling, currency, trivia, poker, slots, AI chat, and more!*
 
 **How does leveling/XP work?**
 
 > You can earn XP by participating on the server in many ways: posting messages, adding reactions, interacting with the bot, and more!
+>  
 > When you level up, you'll earn a random Star Trek-related badge! Check your badges by using the `/badges showcase` command.
-> Check your level, XP, number of badges and more by using the `/profile` command.
+>  
+> Check your level, XP, number of badges and more by using the `/profile display` command.
 
 **What's the difference when I start a message with "computer:" and "agimus:"?**
 
 > The "__computer:__" prefix will tell the bot to try to find factual information about whatever you posted. e.g.: "*computer: convert 24f to celsius*"
+>  
 > The "__agimus:__" prefix will tell the bot to come up with something from its imagination -- using AI algorithms and various public data. e.g.: "*agimus: what is the meaning of life?*"
 
 **The bot don't work!**
